+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-﻿## v3.0.0 - 5 Feb 2024
+﻿## v3.0.1
+* Fix code example 7 and 8 in readme
+* Adjust readme code example sto use filePath instead of filename or path
+
+## v3.0.0 - 5 Feb 2024
 * Switch to .Net 8.0
 * Updated dependencies
 * Improved error handling for empty and whitespace path